Share retain leaf rewrite helpers
Move retain's changed-leaf rewrite state and rebuilt-root finishing into the shared subtree rebuild module. This keeps retain behavior unchanged while making those pieces available to later B-tree mutation paths.
